requirements.txt
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
fastmcp>=2.0.0
